feat(map): added mock posts and refractored some widgets
